Sync minigame_countdown
