+++ include/minigame_countdown.h
@@ -1,0 +1,7 @@
+#ifndef GUARD_MINIGAME_COUNTDOWN_H
+#define GUARD_MINIGAME_COUNTDOWN_H
+
+void StartMinigameCountdown(u16 tilesTag, u16 palTag, s16 x, s16 y, u8 subpriority);
+bool32 IsMinigameCountdownRunning(void);
+
+#endif //GUARD_MINIGAME_COUNTDOWN_H
